+++ core/src/com/game/agar/Game.java
@@ -9,10 +9,14 @@ import com.game.agar.entities.Entity;
 import com.game.agar.rendering.IRenderer;
 import com.game.agar.rendering.SceneRenderer;
 import com.game.agar.tools.Position;
+import com.sun.org.apache.xpath.internal.operations.Bool;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.ArrayList;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Predicate;
 
 
 public class Game extends ApplicationAdapter{
@@ -44,6 +48,7 @@ public class Game extends ApplicationAdapter{
 	@Override
 	public void render () {
 		player.move(1);
+
 		update();
 		camera.setPosition(player.getPosition());
 		renderer.renderFrame();
@@ -55,16 +60,10 @@ public class Game extends ApplicationAdapter{
 	}
 
 
-	public void update() {
-		List <Entity> toDelete = new ArrayList<>();
-		for (Iterator<Entity> it = entities.iterator(); it.hasNext(); ) {
-			Entity currentEntity = it.next();
-			if (!currentEntity.equals(player)) {
-				if (currentEntity.isCollision(player)) {
-					toDelete.add(currentEntity.getEatenEntity(player));
-				}
-			}
-		}
-		entities.removeAll(toDelete);
+	private void update() {
+		Predicate<Entity> isEntityColliding = entity -> !entity.equals(player) && entity.isCollision(player);
+
+		entities.stream().filter(isEntityColliding).forEach(entity -> player.eat(entity));
+		entities.removeIf(isEntityColliding);
 	}
 }
