+++ core/src/com/game/agar/entities/Player.java
@@ -6,9 +6,7 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.game.agar.tools.Position;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
+import java.util.*;
 
 public class Player extends Entity {
 
@@ -26,6 +24,32 @@ public class Player extends Entity {
     public void move(double translation) {
         position.x += translation * Math.cos(movingDirectionAngle);
         position.y += translation * Math.sin(movingDirectionAngle);
+    }
+
+    public void eat (Entity collidingObject){
+        int massGained = collidingObject.getWeight();
+        Timer timer = new Timer();
+        TimerTask growth = new TimerTask() {
+            int targetWeight = weight + massGained ;
+            int gain = 1 + massGained/1000;
+            @Override
+            public void run() {
+                if (weight == targetWeight) {
+                    cancel();
+                    timer.cancel();
+                    timer.purge();
+                }
+                else
+                {
+                    if(weight + gain <=  targetWeight)
+                        weight += gain;
+                    else
+                        weight = targetWeight;
+                    updateRadius();
+                }
+            }
+        };
+        timer.schedule(growth,0,1);
     }
 
 }
