+++ core/src/com/game/agar/rendering/SceneRenderer.java
@@ -34,7 +34,6 @@ public class SceneRenderer implements IRenderer{
         shapeRenderer.setTransformMatrix(camera.getMatrix());
         shapeRenderer.begin(ShapeRenderer.ShapeType.Filled);
         {
-            shapeRenderer.setColor(Color.BLUE);
             entities.forEach(this::renderEntity);
         }
         shapeRenderer.end();
